+++ src/main/java/org/jabref/gui/groups/GroupTreeView.java
@@ -251,6 +251,9 @@ public class GroupTreeView extends BorderPane {
 
         // Filter text field
         setupClearButtonField(searchField);
+
+        // set GroupTree in StateManger
+        this.stateManager.setGroupTree(this.groupTree);
     }
 
     private StackPane getArrowCell(GroupNodeViewModel viewModel) {